+++ R/R/ohlc.R
@@ -3,13 +3,14 @@
 #' @param x \code{xts} object with columns named \code{Open}, \code{High}, \code{Low}, \code{Close}, representing OHLC prices.
 #' @param width integer width of the rolling window to use, or vector of endpoints defining the intervals to use.
 #' @param method one of \code{"O"}, \code{"OC"}, \code{"OHL"}, \code{"OHLC"}, \code{"C"}, \code{"CO"}, \code{"CHL"}, \code{"CHLO"}, or any combination of them, e.g. \code{"OHLC.CHLO"}.
-#' @param na.rm a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds.
+#' @param signed a \code{logical} value indicating whether non-positive estimates should be preceded by the negative sign instead of being imputed. Default \code{FALSE}.
+#' @param na.rm a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds. Default \code{FALSE}.
 #'
 #' @return Time series of spread estimates.
 #'
 #' @keywords internal
 #'
-OHLC <- function(x, width = nrow(x), method = "OHLC.CHLO", na.rm = FALSE){
+OHLC <- function(x, width = nrow(x), method = "OHLC.CHLO", signed = FALSE, na.rm = FALSE){
 
   # methods
   methods <- strsplit(method, split = ".", fixed = TRUE)
@@ -89,15 +90,20 @@ OHLC <- function(x, width = nrow(x), method = "OHLC.CHLO", na.rm = FALSE){
     expr <- sprintf("(%s)/%s", paste0("S2.", m, collapse = "+"), length(m))
     S2 <- cbind(S2, eval(parse(text = expr)))
   }
-
-  # set to zero
-  S2[S2<0] <- 0
-  S2[is.infinite(S2)] <- NA
-
+  
+  # square root
+  S <- sign(S2) * sqrt(abs(S2))
+  
+  # negative estimates
+  if(!signed) S[S<0] <- 0
+  
+  # infinite estimates
+  S[is.infinite(S)] <- NA
+  
   # set names
-  colnames(S2) <- method
+  colnames(S) <- method
 
   # return
-  return(sqrt(S2))
+  return(S)
 
 }