+++ R/R/spread.R
@@ -1,6 +1,6 @@
 #' Estimation of Bid-Ask Spreads from OHLC Prices
 #'
-#' This function estimates bid-ask spreads from open, high, low, and close prices with several methods.
+#' This function estimates (percent) effective bid-ask spreads from open, high, low, and close prices with several methods.
 #'
 #' @details
 #' The method \code{EDGE} implements the Efficient Discrete Generalized Estimator proposed in Ardia-Guidotti-Kroencke (2021).
@@ -11,30 +11,39 @@
 #'
 #' The method \code{AR} implements the estimator proposed in Abdi & Ranaldo (2017). \code{AR2} implements the 2-period adjusted version.
 #'
-#' The method \code{CS} implements the estimator proposed in Corwin & Schultz (2012). \code{CS2} implements the 2-period adjusted version. Both versions are adjusted for overnight returns as described in the paper.
+#' The method \code{CS} implements the estimator proposed in Corwin & Schultz (2012). \code{CS2} implements the 2-period adjusted version. Both versions are adjusted for overnight (close-to-open) returns as described in the paper.
 #'
 #' The method \code{ROLL} implements the estimator proposed in Roll (1984).
 #'
 #' @param x \code{xts} object with columns named \code{Open}, \code{High}, \code{Low}, \code{Close}, representing OHLC prices.
 #' @param width integer width of the rolling window to use, or vector of endpoints defining the intervals to use. By default, the whole time series is used to compute a single spread estimate.
 #' @param method the estimator(s) to use. Choose one or more of: \code{EDGE}, \code{AR}, \code{AR2}, \code{CS}, \code{CS2}, \code{ROLL}, \code{O}, \code{OC}, \code{OHL}, \code{OHLC}, \code{C}, \code{CO}, \code{CHL}, \code{CHLO}, or \code{GMM}. See details.
-#' @param probs vector of probabilities to compute the critical values when the method \code{EDGE} is selected. By default, the critical values at 2.5\% and 97.5\% are computed.
-#' @param na.rm a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds.
+#' @param probs vector of probabilities to compute the critical values when the method \code{EDGE} is selected.
+#' @param signed a \code{logical} value indicating whether non-positive estimates should be preceded by the negative sign instead of being imputed. Default \code{FALSE}.
+#' @param na.rm a \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds. Default \code{FALSE}.
 #'
-#' @return Time series of spread estimates.
+#' @return Time series of (percent) spread estimates.
+#'
+#' @note 
+#' \itemize{
+#' \item Please cite \href{https://www.ssrn.com/abstract=3892335}{Ardia, Guidotti, Kroencke (2021)} 
+#' when using this package in publication. Hint: type \code{citation("bidask")}
+#' \item Place the URL \url{https://github.com/eguidotti/bidask} 
+#' in a footnote when using this package in other online material.
+#' }
 #'
 #' @references
-#' Roll, R. (1984). A simple implicit measure of the effective bid-ask spread in an efficient market. The Journal of Finance, 39 (4), 1127-1139.
-#' \doi{10.1111/j.1540-6261.1984.tb03897.x}
-#'
-#' Corwin, S. A., & Schultz, P. (2012). A simple way to estimate bid-ask spreads from daily high and low prices. The Journal of Finance, 67 (2), 719-760.
-#' \doi{10.1111/j.1540-6261.2012.01729.x}
+#' Ardia, D., Guidotti E., & Kroencke T. A. (2021). Efficient Estimation of Bid-Ask Spreads from Open, High, Low, and Close Prices. 
+#' Available at SSRN: \url{https://www.ssrn.com/abstract=3892335}
 #'
 #' Abdi, F., & Ranaldo, A. (2017). A simple estimation of bid-ask spreads from daily close, high, and low prices. The Review of Financial Studies, 30 (12), 4437-4480.
 #' \doi{10.1093/rfs/hhx084}
 #' 
-#' Ardia, D., Guidotti E., & Kroencke T. A. (2021). Efficient Estimation of Bid-Ask Spreads from Open, High, Low, and Close Prices. 
-#' Available at SSRN: \url{https://www.ssrn.com/abstract=3892335}
+#' Corwin, S. A., & Schultz, P. (2012). A simple way to estimate bid-ask spreads from daily high and low prices. The Journal of Finance, 67 (2), 719-760.
+#' \doi{10.1111/j.1540-6261.2012.01729.x}
+#' 
+#' Roll, R. (1984). A simple implicit measure of the effective bid-ask spread in an efficient market. The Journal of Finance, 39 (4), 1127-1139.
+#' \doi{10.1111/j.1540-6261.1984.tb03897.x}
 #'
 #' @examples
 #' # simulate a price process with spread 1%
@@ -61,7 +70,7 @@
 #'
 #' @export
 #'
-spread <- function(x, width = nrow(x), method = "EDGE", probs = NULL, na.rm = FALSE){
+spread <- function(x, width = nrow(x), method = "EDGE", probs = NULL, signed = FALSE, na.rm = FALSE){
 
   if(!is.xts(x))
     stop("x must be a xts object")
@@ -74,38 +83,38 @@ spread <- function(x, width = nrow(x), method = "EDGE", probs = NULL, na.rm = FA
 
   m <- "EDGE"
   if(m %in% method){
-    S <- cbind(S, EDGE(x, width = width, probs = probs, na.rm = na.rm))
+    S <- cbind(S, EDGE(x, width = width, probs = probs, signed = signed, na.rm = na.rm))
     method <- setdiff(method, m)
   }
   
   m <- "GMM"
   if(m %in% method){
-    S <- cbind(S, GMM(x, width = width, na.rm = na.rm))
+    S <- cbind(S, GMM(x, width = width, signed = signed, na.rm = na.rm))
     method <- setdiff(method, m)
   }
 
   m <- c("AR","AR2")
   if(any(m %in% method)){
     m <- intersect(method, m)
-    S <- cbind(S, AR(x, width = width, method = m, na.rm = na.rm))
+    S <- cbind(S, AR(x, width = width, method = m, signed = signed, na.rm = na.rm))
     method <- setdiff(method, m)
   }
 
   m <- c("CS","CS2")
   if(any(m %in% method)){
     m <- intersect(method, m)
-    S <- cbind(S, CS(x, width = width, method = m, na.rm = na.rm))
+    S <- cbind(S, CS(x, width = width, method = m, signed = signed, na.rm = na.rm))
     method <- setdiff(method, m)
   }
 
   m <- "ROLL"
   if(m %in% method){
-    S <- cbind(S, ROLL(x, width = width, na.rm = na.rm))
+    S <- cbind(S, ROLL(x, width = width, signed = signed, na.rm = na.rm))
     method <- setdiff(method, m)
   }
   
   if(length(method)){
-    S <- cbind(S, OHLC(x, width = width, method = method, na.rm = na.rm))
+    S <- cbind(S, OHLC(x, width = width, method = method, signed = signed, na.rm = na.rm))
   }
 
   return(S)